requirements.txt
colorlog==6.9.0
docker==7.1.0
pytest==8.3.5
pyyaml==6.0.2